+++ vendor_download.sh
@@ -17,8 +17,8 @@ if [ -z "$GITHUB_TOKEN" ]; then
   exit 1
 fi
 
-rm -f vendor/*
-mkdir -p vendor
+rm -f vendor-bins/*
+mkdir -p vendor-bins
 
 ASSET_POSTFIX=""
 WIGGINS_ASSET_POSTFIX=""
@@ -48,7 +48,7 @@ echo "Downloading asset information from latest tag for architecture '$ASSET_POS
 WIGGINS_TAG="2021-10-21-8eb009f"
 echo "Downloading wiggins binary"
 echo "Using wiggins release: $WIGGINS_TAG"
-WIGGINS_RELEASE_JSON=vendor/wiggins-release.json
+WIGGINS_RELEASE_JSON=vendor-bins/wiggins-release.json
 curl -sSL \
     -H "Authorization: token $GITHUB_TOKEN" \
     -H "Accept: application/vnd.github.v3.raw" \
@@ -59,7 +59,7 @@ FILTER=".name == \"scotland_yard-wiggins-$WIGGINS_ASSET_POSTFIX\""
 jq -c ".assets | map({url: .url, name: .name}) | map(select($FILTER)) | .[]" $WIGGINS_RELEASE_JSON | while read ASSET; do
   URL="$(echo $ASSET | jq -c -r '.url')"
   NAME="$(echo $ASSET | jq -c -r '.name')"
-  OUTPUT="$(echo vendor/$NAME | sed 's/scotland_yard-//' | sed 's/-'$WIGGINS_ASSET_POSTFIX'$//')"
+  OUTPUT="$(echo vendor-bins/$NAME | sed 's/scotland_yard-//' | sed 's/-'$WIGGINS_ASSET_POSTFIX'$//')"
 
   echo "Downloading '$NAME' to '$OUTPUT'"
   curl -sL -H "Authorization: token $GITHUB_TOKEN" -H "Accept: application/octet-stream" -s $URL > $OUTPUT
@@ -70,10 +70,10 @@ echo
 
 if $OS_WINDOWS; then
   echo "Skipping syft for Windows builds"
-  touch vendor/syft
+  touch vendor-bins/syft
 else
   echo "Downloading forked syft binary"
-  SYFT_RELEASE_JSON=vendor/syft-release.json
+  SYFT_RELEASE_JSON=vendor-bins/syft-release.json
   curl -sSL \
       -H "Authorization: token $GITHUB_TOKEN" \
       -H "Accept: application/vnd.github.v3.raw" \
@@ -87,13 +87,13 @@ else
   jq -c ".assets | map({url: .url, name: .name}) | map(select($FILTER)) | .[]" $SYFT_RELEASE_JSON | while read ASSET; do
     URL="$(echo $ASSET | jq -c -r '.url')"
     NAME="$(echo $ASSET | jq -c -r '.name')"
-    OUTPUT=vendor/${NAME%"-$ASSET_POSTFIX"}
+    OUTPUT=vendor-bins/${NAME%"-$ASSET_POSTFIX"}
 
     echo "Downloading '$NAME' to '$OUTPUT'"
     curl -sL -H "Authorization: token $GITHUB_TOKEN" -H "Accept: application/octet-stream" -s $URL > $OUTPUT
     echo "Extracting syft binary from tarball"
     tar xzf $OUTPUT fossa-container-scanning
-    mv fossa-container-scanning vendor/syft
+    mv fossa-container-scanning vendor-bins/syft
     rm $OUTPUT
 
   done
@@ -103,11 +103,11 @@ fi
 
 if $OS_WINDOWS; then
   echo "Skipping cliv1 for Windows builds"
-  touch vendor/cliv1
+  touch vendor-bins/cliv1
 else
   echo ""
   echo "Downloading cliv1 binary"
-  CLIV1_RELEASE_JSON=vendor/cliv1-release.json
+  CLIV1_RELEASE_JSON=vendor-bins/cliv1-release.json
   curl -sSL \
       -H "Authorization: token $GITHUB_TOKEN" \
       -H "Accept: application/vnd.github.v3.raw" \
@@ -121,13 +121,13 @@ else
   jq -c ".assets | map({url: .url, name: .name}) | map(select($FILTER)) | .[]" $CLIV1_RELEASE_JSON | while read ASSET; do
     URL="$(echo $ASSET | jq -c -r '.url')"
     NAME="$(echo $ASSET | jq -c -r '.name')"
-    OUTPUT=vendor/${NAME%"-$ASSET_POSTFIX"}
+    OUTPUT=vendor-bins/${NAME%"-$ASSET_POSTFIX"}
 
     echo "Downloading '$NAME' to '$OUTPUT'"
     curl -sL -H "Authorization: token $GITHUB_TOKEN" -H "Accept: application/octet-stream" -s $URL > $OUTPUT
     echo "Extracting cliv1 binary from tarball"
     tar xzf $OUTPUT fossa
-    mv fossa vendor/cliv1
+    mv fossa vendor-bins/cliv1
     rm $OUTPUT
 
   done
@@ -136,10 +136,10 @@ else
 fi
 
 echo "Marking binaries executable"
-chmod +x vendor/*
+chmod +x vendor-bins/*
 
 echo "Compressing binaries"
-upx vendor/* || echo "WARN: 'upx' command not found, binaries will not be compressed"
+upx vendor-bins/* || echo "WARN: 'upx' command not found, binaries will not be compressed"
 
 echo "Vendored binaries are ready for use"
-ls -lh vendor/
+ls -lh vendor-bins/