+++ src/System/CGroup.hs
@@ -1,0 +1,6 @@
+module System.CGroup (
+  module X,
+) where
+
+import System.CGroup.CPU as X
+import System.CGroup.Types as X